% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_add_line_time}
\alias{its_add_line_time}
\title{Add another line to an existing plot}
\usage{
its_add_line_time(a = 1, b = 0, colour = "dodgerblue")
}
\arguments{
\item{a}{the slope of the new line}

\item{b}{the intercept of the new line}

\item{colour}{colour of the new line}
}
\value{
a ggplot2 geom
}
\description{
Add another line to an existing plot
}
\examples{
its_axplusb_time() + its_add_line_time(a=4, colour="orange")

}
