% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_axplusb_time}
\alias{its_axplusb_time}
\title{Draw a straight line graph}
\usage{
its_axplusb_time(xrange = c(-4, 4), yrange = NULL, a = 1, b = 0)
}
\arguments{
\item{xrange}{the extent of the x-axis}

\item{yrange}{the (optional) extent of the y-axis}

\item{a}{the slope of the line}

\item{b}{the intercept of the line}
}
\value{
a ggplot2 plot
}
\description{
Draw a straight line graph
}
\examples{
its_axplusb_time()
its_axplusb_time(xrange = c(-10, 10), a = 4, b = 3)
}
