% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_categoric_scatter_time}
\alias{its_categoric_scatter_time}
\title{returns a categoric scatter plot}
\usage{
its_categoric_scatter_time(
  df,
  names_to = "group",
  values_to = "value",
  colour = "dodgerblue",
  join_tops = FALSE
)
}
\description{
returns a categoric scatter plot
}
