% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_hot_dog_and_ice_cream_plot_time}
\alias{its_hot_dog_and_ice_cream_plot_time}
\title{hot dog and ice cream box plots}
\usage{
its_hot_dog_and_ice_cream_plot_time()
}
\description{
hot dog and ice cream box plots
}
