% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataobjs.R
\name{its_hot_dog_and_ice_cream_time}
\alias{its_hot_dog_and_ice_cream_time}
\title{interaction data from Jim's stats site}
\usage{
its_hot_dog_and_ice_cream_time()
}
\description{
interaction data from Jim's stats site
}
