% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_hot_dog_and_ice_cream_two_ways_time}
\alias{its_hot_dog_and_ice_cream_two_ways_time}
\title{plot interaction in food and condiment}
\usage{
its_hot_dog_and_ice_cream_two_ways_time()
}
\description{
plot interaction in food and condiment
}
