% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_is_the_mean_a_good_summary_time}
\alias{its_is_the_mean_a_good_summary_time}
\title{plots of goodness of the mean}
\usage{
its_is_the_mean_a_good_summary_time(n, type = "hist")
}
\arguments{
\item{n}{number of points to generate}

\item{type}{hist or jitter type of plot to return}
}
\description{
plots of goodness of the mean
}
