% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_multi_category_with_lines_time}
\alias{its_multi_category_with_lines_time}
\title{returns a multi categry line plot}
\usage{
its_multi_category_with_lines_time()
}
\description{
returns a multi categry line plot
}
