% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_plot_xy_time}
\alias{its_plot_xy_time}
\title{plot xy data}
\usage{
its_plot_xy_time(df, line = FALSE, residuals = FALSE)
}
\arguments{
\item{df}{dataframe with columns x and y}

\item{line}{draw the computed line}

\item{residuals}{draw the residuals}
}
\value{
ggplot
}
\description{
plot xy data
}
