% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_random_xy_time}
\alias{its_random_xy_time}
\title{Generate some x and y data in a dataframe}
\usage{
its_random_xy_time(n, min = 5, max = 15, mult = 2, seed = 456)
}
\arguments{
\item{n}{number of points}

\item{min}{minimum random value}

\item{max}{maximum random value}

\item{mult}{how much bigger on average y is than x}

\item{seed}{random seed}
}
\description{
Generate some x and y data in a dataframe
}
