% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataobjs.R
\name{its_remove_a_group_time}
\alias{its_remove_a_group_time}
\title{removes a group from a dataframe}
\usage{
its_remove_a_group_time(df, col = "group", level = "trt2")
}
\description{
removes a group from a dataframe
}
