% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_summary_plot_time}
\alias{its_summary_plot_time}
\title{returns a plot summarising hypothesis tests as a figure}
\usage{
its_summary_plot_time()
}
\description{
returns a plot summarising hypothesis tests as a figure
}
