% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{its_wide_to_long_time}
\alias{its_wide_to_long_time}
\title{returns a long version of a table}
\usage{
its_wide_to_long_time(df, names_to = "group", values_to = "value")
}
\description{
returns a long version of a table
}
